Factor out server function to dedicated module
